fixed local files example
